demo/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
open-clip-torch>=2.26.1
torch>=2.2.0
Pillow>=10.0.0
numpy>=1.26.0
runpod>=1.6.0
